--- src/main/scala/io/iohk/ethereum/network/p2p/PV62.scala
@@ -1,5 +1,0 @@
-package io.iohk.ethereum.network.p2p
-
-package object PV62 {
-
-}
